+++ lib/ui/screens/settings/settings_button_info.dart
@@ -1,0 +1,11 @@
+class ButtonInfo {
+  final bool active;
+  final String title;
+  final Function() action;
+
+  ButtonInfo({
+    required this.active,
+    required this.title,
+    required this.action,
+  });
+}
